requirements.txt
django 
djangorestframework 
gunicorn
Pillow